+++ src/components/input/InputTexture.vue
@@ -3,12 +3,14 @@
 		<div
 			class="rounded border border-ui-menu-outline bg-ui-menu-inner px-1 py-0.5 text-ui-menu-text
 				grid grid-cols-[1fr_auto] gap-1"
-			tabindex="0"
+			:tabindex="disabled ? undefined : 0"
 		>
 			<span class="overflow-hidden text-ellipsis" :title="model?.name">
 				{{ model?.name || 'None' }}
 			</span>
-			<button v-if="model" type="button" class="cursor-pointer px-1" @click="reset">x</button>
+			<button v-if="model" type="button" class="cursor-pointer px-1" @click="reset">
+				<MxIcon name="ui/close" />
+			</button>
 		</div>
 		<div class="hidden grid-cols-2 group-focus-within:grid">
 			<button type="button" class="btn" @click="openLibrary">Library</button>
@@ -24,6 +26,10 @@ import { loadEXR } from '@/three/modules/loaders/exr'
 import { loadTexture } from '@/three/modules/loaders/textureLoader'
 import { isPolyHavenFileInfo } from '@/utils/polyhaven'
 import { useFileDialog } from '@vueuse/core'
+
+defineProps<{
+	disabled?: boolean
+}>()
 
 const model = defineModel<THREE.Texture | null>()